requirements.txt
fastapi==0.115.2
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
pydantic==2.10.4